refactor(worker): one file per group of agent tools
agent-worker-tools.ts was 1634 lines holding fifteen unrelated tool-group
factories, so reading any one of them meant scrolling past all the others and
every new tool grew the same file. Each group now has its own file and the
original keeps only the barrel the rest of the worker imports, unchanged, so no
caller moves.

The shared pieces the groups all reached for — the timeout wrapper, the result
truncation and readers, the card schemas, the taint and gate types — sit in
tool-primitives, which none of the groups export back into.

Nothing is rewritten. Every line is where it was, only in a smaller file: the
concatenation of the new files matches the original exactly once imports and
export keywords are set aside, which is how this was checked. GateOutcome had
no importer and is no longer exported.
